+++ CHANGELOG.md
@@ -10,6 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - IPAddress
   - Improved integration test structure.
 
+### Added
+
+- NetIPInterface
+  - Added `InterfaceMetric` parameter- fixes [Issue #473](https://github.com/PowerShell/xNetworking/issues/473).
+
 ### Fixed
 
 - NetIPInterface